requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
async-timeout==5.0.1
attrs==24.3.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
databases==0.9.0
distro==1.9.0
Django==4.2.19
django-bootstrap-v5==1.0.11
django-tables2==2.7.3
djangorestframework==3.15.2
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
fonttools==4.55.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==6.5.2
inflection==0.5.1
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
multidict==6.1.0
nltk==3.9.1
numpy==2.0.2
openai==0.28.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.5
pydantic-core==2.27.2
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.4.1
SQLAlchemy==2.0.38
sqlparse==0.4.4
starlette==0.45.3
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.1
typing-extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wordcloud==1.9.4
yarl==1.18.3
zipp==3.15.0